Bug fix: TimeOutException causing extra call to reward
After a TimeOutException, we would always call the reward function on
a randomly selected arm. But the timeout may have occurred after
calling the reward already once for the current batch (during
cp.learn for instance), which would mean that we would call the
reward function twice for the same batch. As a consequence, this bug
gave a few extra points to click predictors that would produce timeout
exceptions.